Extract config into own module.
